fixed turn number attention for context = 0
