+++ allennlp/models/reading_comprehension/__init__.py
@@ -7,5 +7,5 @@ These submodules contain models for things that are predominantly focused on rea
 
 from allennlp.models.reading_comprehension.bidaf import BidirectionalAttentionFlow
 from allennlp.models.reading_comprehension.bidaf_ensemble import BidafEnsemble
-from allennlp.models.reading_comprehension.dqa import DQA
+from allennlp.models.reading_comprehension.dialog_qa import DialogQA
