+++ allennlp/modules/__init__.py
@@ -19,8 +19,6 @@ from allennlp.modules.span_pruner import SpanPruner
 from allennlp.modules.text_field_embedders import TextFieldEmbedder
 from allennlp.modules.time_distributed import TimeDistributed
 from allennlp.modules.token_embedders import TokenEmbedder, Embedding
-from allennlp.modules.tri_linear_attention import  TriLinearAttention
-from allennlp.modules.variational_dropout import VariationalDropout
 from allennlp.modules.matrix_attention import MatrixAttention
 from allennlp.modules.attention import Attention
 from allennlp.modules.input_variational_dropout import InputVariationalDropout
