+++ allennlp/tests/fixtures/dialog_qa/experiment.json
@@ -93,7 +93,6 @@
     "num_epochs": 30,
     "patience": 10,
     "validation_metric": "+f1",
-    "cuda_device": 0, 
     "learning_rate_scheduler":  {
       "type": "reduce_on_plateau",
       "factor": 0.5,
